+++ packages/esrp-npm-release/src/__tests__/generateJwsToken.test.ts
@@ -1,6 +1,6 @@
 import { beforeAll, describe, expect, it } from '@jest/globals';
 import jws from 'jws';
-import { generateJwsToken } from '../utils/generateJwsToken.ts';
+import { generateJwsToken } from '../auth/generateJwsToken.ts';
 import { generateTestCert, isOpensslAvailable, type TestCert } from '../__fixtures__/testCert.ts';
 
 // eslint-disable-next-line no-restricted-properties -- intentional skip when openssl is unavailable
