+++ packages/esrp-npm-release/src/__tests__/npmRelease.test.ts
@@ -2,8 +2,8 @@ import { beforeAll, describe, expect, it } from '@jest/globals';
 import crypto from 'crypto';
 import fs from 'fs';
 import path from 'path';
-import { createNpmReleaseRequest, redactReleaseRequest } from '../models/npmRelease.ts';
-import { FileHashType, type ReleaseFileInfo, type ReleaseRequestMessage } from '../models/types.ts';
+import { createNpmReleaseRequest, redactReleaseRequest } from '../esrpApi/npmRelease.ts';
+import { FileHashType, type ReleaseFileInfo, type ReleaseRequestMessage } from '../esrpApi/types.ts';
 import { generateTestCert, isOpensslAvailable, type TestCert } from '../__fixtures__/testCert.ts';
 import { setupTempDir } from '../__fixtures__/tempDir.ts';
 import { ReleaseError } from '../utils/ReleaseError.ts';
